+++ weaviate/classes/query.py
@@ -1,5 +1,5 @@
 from weaviate.collections.classes.aggregate import Metrics
-from weaviate.collections.classes.filters import Filter
+from weaviate.collections.classes.filters import Filter, FilterReturn
 from weaviate.collections.classes.generative import GenerativeConfig
 from weaviate.collections.classes.grpc import (
     BM25OperatorFactory as BM25Operator,
@@ -24,6 +24,7 @@ from weaviate.collections.classes.types import GeoCoordinate
 __all__ = [
     "Diversity",
     "Filter",
+    "FilterReturn",
     "GeoCoordinate",
     "GenerativeConfig",
     "GroupBy",