+++ weaviate/collections/classes/config_named_vectors.py
@@ -14,6 +14,7 @@ from weaviate.collections.classes.config_vector_index import (
     _VectorIndexConfigCreate,
     _VectorIndexConfigDynamicUpdate,
     _VectorIndexConfigFlatUpdate,
+    _VectorIndexConfigHFreshUpdate,
     _VectorIndexConfigHNSWUpdate,
     _VectorIndexConfigUpdate,
 )
@@ -590,6 +591,7 @@ This method is deprecated and will be removed in Q2 '25. Please use :meth:`~weav
             vectorizer=_Multi2VecGoogleConfig(
                 projectId=project_id,
                 location=location,
+                audioFields=None,
                 imageFields=_map_multi2vec_fields(image_fields),
                 textFields=_map_multi2vec_fields(text_fields),
                 videoFields=_map_multi2vec_fields(video_fields),
@@ -606,6 +608,7 @@ This method is deprecated and will be removed in Q2 '25. Please use :meth:`~weav
         *,
         location: str,
         project_id: str,
+        audio_fields: Optional[Union[List[str], List[Multi2VecField]]] = None,
         image_fields: Optional[Union[List[str], List[Multi2VecField]]] = None,
         text_fields: Optional[Union[List[str], List[Multi2VecField]]] = None,
         video_fields: Optional[Union[List[str], List[Multi2VecField]]] = None,
@@ -626,6 +629,7 @@ This method is deprecated and will be removed in Q2 '25. Please use :meth:`~weav
             vectorize_collection_name: Whether to vectorize the collection name. Defaults to `True`.
             location: Where the model runs. REQUIRED.
             project_id: The project ID to use, REQUIRED.
+            audio_fields: The audio fields to use in vectorization.
             image_fields: The image fields to use in vectorization.
             text_fields: The text fields to use in vectorization.
             video_fields: The video fields to use in vectorization.
@@ -638,6 +642,7 @@ This method is deprecated and will be removed in Q2 '25. Please use :meth:`~weav
             vectorizer=_Multi2VecGoogleConfig(
                 projectId=project_id,
                 location=location,
+                audioFields=_map_multi2vec_fields(audio_fields),
                 imageFields=_map_multi2vec_fields(image_fields),
                 textFields=_map_multi2vec_fields(text_fields),
                 videoFields=_map_multi2vec_fields(video_fields),
@@ -1340,6 +1345,7 @@ class _NamedVectorsUpdate:
         *,
         vector_index_config: Union[
             _VectorIndexConfigHNSWUpdate,
+            _VectorIndexConfigHFreshUpdate,
             _VectorIndexConfigFlatUpdate,
             _VectorIndexConfigDynamicUpdate,
         ],